account/requirements.txt
APScheduler==3.10.4
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
py-request-nacos==0.0.11
PyMySQL==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
Werkzeug==3.0.0
zipp==3.17.0
